+++ GHEtool/Examples/custom_borefield_configuration.py
@@ -38,12 +38,11 @@ def custom_borefield_configuration():
     monthly_load_heating = annual_heating_load * monthly_load_heating_percentage  # kWh
     monthly_load_cooling = annual_cooling_load * monthly_load_cooling_percentage  # kWh
 
-    # create the borefield object
+    # set the load
+    load = MonthlyGeothermalLoadAbsolute(monthly_load_heating, monthly_load_cooling, peak_heating, peak_cooling)
 
-    borefield = Borefield(peak_heating=peak_heating,
-                          peak_cooling=peak_cooling,
-                          baseload_heating=monthly_load_heating,
-                          baseload_cooling=monthly_load_cooling)
+    # create the borefield object
+    borefield = Borefield(load=load)
 
     borefield.set_ground_parameters(data)
     borefield.Rb = 0.2