+++ GHEtool/Examples/effect_of_borehole_configuration.py
@@ -3,7 +3,7 @@ This document is an example of how the borefield configuration can influence the
 """
 
 # import all the relevant functions
-from GHEtool import GroundConstantTemperature, Borefield
+from GHEtool import GroundConstantTemperature, Borefield, MonthlyGeothermalLoadAbsolute
 import numpy as np
 import pygfunction as gt
 
@@ -29,12 +29,11 @@ def effect_borefield_configuration():
     monthly_load_heating = annual_heating_load * monthly_load_heating_percentage   # kWh
     monthly_load_cooling = annual_cooling_load * monthly_load_cooling_percentage   # kWh
 
-    # create the borefield object
+    # set the load
+    load = MonthlyGeothermalLoadAbsolute(monthly_load_heating, monthly_load_cooling, peak_heating, peak_cooling)
 
-    borefield = Borefield(peak_heating=peak_heating,
-                          peak_cooling=peak_cooling,
-                          baseload_heating=monthly_load_heating,
-                          baseload_cooling=monthly_load_cooling)
+    # create the borefield object
+    borefield = Borefield(load=load)
 
     borefield.set_ground_parameters(data)
     borefield.set_borefield(borefield_gt)